Reorganize .structure submodule
